refactor edit connector page for smaller file size
